Reformat eBayDataManager to exclude fetching token - single responsibility principle
