+++ C7FIT/eBayDataManager.swift
@@ -8,73 +8,20 @@
 
 import Foundation
 
-class eBayDataManager {
+struct eBayDataManager {
     
     // MARK: - Constants
     
-    let encodedOAuthCreds = "<INSERT CREDENTIALS HERE>"
-    let clientKey = "<INSERT CREDENTIALS HERE>"
-    let authorizeURL = "https://api.sandbox.ebay.com/identity/v1/oauth2/token"
-    let eBayRedirectURLName = "<INSERT CREDENTIALS HERE>"
-    
     let browseAPIbaseURL = "https://api.ebay.com/buy/browse/v1/"
     
-    // MARK: - Properties
-    
-    var OAuth2Token: String = ""
-    
-    // MARK: - Initialization
-    
-    // TODO: A bit hacky... Find a better implementation of mutating self parameters if possible
-    init() {
-        self.getOAuth2Token() { token in
-            guard let token = token else { return }
-            self.OAuth2Token = token
-        }
-    }
-    
-    /**
-        Obtain a new bearer token
-        - Returns completion: A callback that returns an optional string token
-     */
-    func getOAuth2Token(completion: @escaping (_:String?) -> Void) {
-        let headers = [
-            "content-type": "application/x-www-form-urlencoded",
-            "authorization": "Basic \(encodedOAuthCreds)",
-            "cache-control": "no-cache"
-        ]
-        let url = URL(string: "\(authorizeURL)?grant_type=client_credentials&redirect_uri=https%3A%2F%2Fsignin.ebay.com%2Fauthorize%3Fclient_id%3D\(clientKey)%26response_type%3Dcode%26redirect_uri%3D\(eBayRedirectURLName)%26scope%3Dhttps%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fbuy.order.readonly%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fbuy.guest.order%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.marketing.readonly%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.marketing%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.inventory.readonly%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.inventory%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.account.readonly%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.account%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.fulfillment.readonly%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.fulfillment%20https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope%2Fsell.analytics.readonly&scope=https%3A%2F%2Fapi.ebay.com%2Foauth%2Fapi_scope")!
-        
-        let request = NSMutableURLRequest(url: url, cachePolicy: .useProtocolCachePolicy, timeoutInterval: 10.0)
-        request.httpMethod = "POST"
-        request.allHTTPHeaderFields = headers
-        
-        let dataTask = URLSession.shared.downloadTask(with: request as URLRequest) { (location, response, error) in
-            guard let location = location, error == nil else {
-                print("Error in retrieving bearer token: \(error?.localizedDescription)")
-                return
-            }
-            guard let dataObject = try? Data(contentsOf: location), let dataJSON = try? JSONSerialization.jsonObject(with: dataObject, options: []) else {
-                return
-            }
-            guard let dataDict = dataJSON as? [String: Any], let token = dataDict["access_token"] as? String else {
-                return
-            }
-            DispatchQueue.main.async {
-                completion("Bearer \"\(token)\"")
-            }
-        }
-        
-        dataTask.resume()
-    }
-  
+    // MARK: - Network Requests
     
     /**
         Fetches an item based off of its ID.
          - Parameter itemID: ItemID
          - Returns completion: A callback that returns the item JSON
      */
-    func getItem(itemID: String, completion: @escaping ([String: Any]?) -> Void) {
+    func getItem(itemID: String, OAuth2Token: String, completion: @escaping ([String: Any]?) -> Void) {
         let headers = [
             "authorization": OAuth2Token
         ]
@@ -105,9 +52,9 @@ class eBayDataManager {
         - Parameter query: The item search query
         - Returns completion: A callback that returns an eBayItemCategory model
      */
-    func searchItem(query: String, completion: @escaping(eBayItemCategory) -> Void) {
+    func searchItem(query: String, OAuth2Token: String, completion: @escaping(eBayItemCategory) -> Void) {
         let headers = [
-            "authorization": self.OAuth2Token
+            "authorization": OAuth2Token
         ]
         let urlEncodedString = "\(self.browseAPIbaseURL)item_summary/search?q=\(query)&limit=10".addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed)!
         let url = URL(string: urlEncodedString)!
